requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
boto3==1.34.151
botocore==1.34.151
certifi==2024.7.4
charset-normalizer==3.3.2
Django>=5.0, <5.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
packaging==24.1
pillow==10.4.0
psycopg2-binary
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
replicate==0.31.0
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
dj-database-url
stripe==9.8.0